+++ src/app/components/molecules/footer/FooterComponent.tsx
@@ -2,7 +2,7 @@ import styles from './footer.module.css';
 import Image from 'next/image';
 
 import footerlogo from '@images/footer-logo.svg';
-import InviteButton from '@components/atoms/inviteButton/InviteButtonComponent';
+import InviteButton from '@components/atoms/invite-button/InviteButtonComponent';
 import LinkList from '@components/atoms/link-list/LinkList';
 import IconBar from '@components/atoms/icon-bar/IconBar';
