+++ src/app/page.module.css
@@ -3,7 +3,6 @@
   flex-direction: column;
   align-items: center;
   /* padding: 6rem; */
-  min-height: 100vh;
   background-color: white;
 
   caret-color: transparent;
