+++ src/app/page.tsx
@@ -1,9 +1,9 @@
 import { Inter } from 'next/font/google';
 import styles from './page.module.css';
 import Navbar from '@components/molecules/navbar/NavbarComponent';
-import HeroSection from '@components/organisms/herosec/HeroSectionOrg';
-import WhySection from '@components/organisms/whySec/WhySectionOrg';
-import LatestSection from '@components/organisms/latestSec/LatestArticlesOrg';
+import HeroSection from '@components/organisms/hero-section/HeroSectionOrg';
+import WhySection from '@components/organisms/why-section/WhySectionOrg';
+import LatestSection from '@components/organisms/latest-section/LatestArticlesOrg';
 import Footer from '@components/molecules/footer/FooterComponent';
 
 const inter = Inter({
